requirements.txt
timm
torch
torchvision
facenet-pytorch
opencv-python
Pillow
PyYAML
numpy
